requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.7.0
pandas>=2.0.0
seaborn>=0.12.0
roboflow>=1.1.0
PyYAML>=6.0
tqdm>=4.65.0
Pillow>=9.5.0
split-folders>=0.5.1
yt-dlp>=2024.0.0